Update to latest Kotlin 2.2.0
