+++ build-logic/plugins/src/main/kotlin/com/amplify/ui/ConfigureAndroid.kt
@@ -1,0 +1,67 @@
+/*
+ * Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+package com.amplify.ui
+
+import com.android.build.api.dsl.CommonExtension
+import org.gradle.api.Project
+
+internal fun Project.configureAndroid(extension: CommonExtension<*, *, *, *, *, *>) {
+    val sdkVersionName = findProperty("VERSION_NAME") ?: rootProject.findProperty("VERSION_NAME")
+
+    extension.apply {
+        compileSdk = 35
+
+        buildFeatures {
+            buildConfig = true
+        }
+
+        defaultConfig {
+            minSdk = 24
+            testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+            testInstrumentationRunnerArguments += "clearPackageData" to "true"
+
+            testOptions {
+                animationsDisabled = true
+                unitTests {
+                    isIncludeAndroidResources = true
+                }
+            }
+
+            buildConfigField("String", "VERSION_NAME", "\"$sdkVersionName\"")
+        }
+
+        lint {
+            warningsAsErrors = true
+            abortOnError = true
+            enable += listOf("UnusedResources")
+            disable += listOf("GradleDependency", "NewerVersionAvailable", "AndroidGradlePluginVersion")
+        }
+
+        // Needed when running integration tests. The oauth2 library uses relies on two
+        // dependencies (Apache's httpcore and httpclient), both of which include
+        // META-INF/DEPENDENCIES. Tried a couple other options to no avail.
+        packaging {
+            resources.excludes += setOf("META-INF/DEPENDENCIES", "META-INF/LICENSE*")
+        }
+
+        buildFeatures {
+            compose = true
+        }
+
+        composeOptions {
+            kotlinCompilerExtensionVersion = "1.5.3"
+        }
+    }
+}
